Create runner.go file
